# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Parkly API Documentation",
        "description": "API documentation for Parkly backend",
        "contact": {
            "email": "stefan.pejchinoski@iwconnect.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://16.171.224.141",
            "description": "API Server"
        },
        {
            "url": "/"
        }
    ],
    "paths": {
        "/api/facilities": {
            "get": {
                "tags": [
                    "Facilities"
                ],
                "summary": "Get a list of all facilities",
                "description": "Display a listing of facilities.",
                "operationId": "11eafb078d1797dd2d2a2a8459ead7f9",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Downtown Parking"
                                                    },
                                                    "parking_spot_count": {
                                                        "type": "integer",
                                                        "example": 50
                                                    },
                                                    "manager_id": {
                                                        "type": "integer",
                                                        "example": 5,
                                                        "nullable": true
                                                    },
                                                    "manager": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 5
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "John Manager"
                                                            },
                                                            "email": {
                                                                "type": "string",
                                                                "example": "manager@example.com"
                                                            }
                                                        },
                                                        "type": "object",
                                                        "nullable": true
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Facilities"
                ],
                "summary": "Create a new facility",
                "description": "Store a newly created facility.",
                "operationId": "9a393251624df76dbe1b3b8e8ff4472d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "parking_spot_count"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "East Side Parking"
                                    },
                                    "parking_spot_count": {
                                        "type": "integer",
                                        "example": 25
                                    },
                                    "manager_id": {
                                        "type": "integer",
                                        "example": 5,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Facility created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "East Side Parking"
                                                },
                                                "parking_spot_count": {
                                                    "type": "integer",
                                                    "example": 25
                                                },
                                                "manager_id": {
                                                    "type": "integer",
                                                    "example": 5,
                                                    "nullable": true
                                                },
                                                "manager": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 5
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "John Manager"
                                                        },
                                                        "email": {
                                                            "type": "string",
                                                            "example": "manager@example.com"
                                                        }
                                                    },
                                                    "type": "object",
                                                    "nullable": true
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "name": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The name field is required."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - User doesn't have permission"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}